ml-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.3.0+cpu
torchvision==0.18.0+cpu
pillow==10.3.0
requests==2.31.0
pydantic==2.7.1
huggingface_hub==0.23.0